# RavenDB-19730 Querying performance / memory improvements