… src/Raven.Server/Documents/Sharding/Commands/Querying/AbstractShardedQueryCommand.cs