### Issue link

https://issues.hibernatingrhinos.com/issue/RavenDB-19730/Sharding-Query-Further-Performance-Optimizations

### Additional description

- Avoid materializing blittable query to string - it's read only so each shard request executor can read from it independently (each async blittable writer uses own context for writing) - this requered to pass external context to GetPropertiesByInsertionOrder
- Instead of creating a blittable from InsexQueryServerSize instance we can reuse the original JSON that we read from the request - SourceQueryJson
- Making sure that request sent to shards aren't gzip compressed
- Removing unnecessary abstraction - AbstractShardedQueryCommand

- Continuation work of https://github.com/ravendb/ravendb/pull/15851

### Testing results:

#### Local machine

- Query: `from "Questions" where search("Title",$p0) select Title limit 128` // $p0: {hot-term} 
- Duration: 30 seconds
- Concurrency: 50 threads

#### Non sharded

Client side stats:

<html><body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
NonSharded | 2072 | 17 | 114 | 169 | 197 | 50 | 2610 | 62201

<!--EndFragment-->
</body>
</html>

Server side stats:

![image](https://user-images.githubusercontent.com/376551/225014140-3eb77284-739b-46b5-b5ac-a018e9cbe57c.png)

#### Sharded - 2 nodes, A - orchestrator, B - 3 shards (everything on the same machine)


Client side stats:

<html><body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Before changes |   |   |   |   |   |   |   |  
Orch seperated | 580 | 81 | 366 | 378 | 2678 | 309 | 2793 | 17454
After changes |   |   |   |   |   |   |   |  
Orch seperated | 717 | 64 | 296 | 315 | 2613 | 365 | 2731 | 21545

<!--EndFragment-->
</body>
</html>

Server side stats (orchestrator):

![image](https://user-images.githubusercontent.com/376551/225014743-bae71bab-fe53-48a6-bade-37e8741a13a9.png)


#### Cluster in RavenDB Cloud

Setup:

- client running queries - separate machine in the same AWS region as RavenDB servers, SSL connection via public host name
- 3 nodes cluster - A - orchestrator, B and C - 2 shards (everything on the same machine)

<html><body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
   NonSharded    | 286 | 170 | 591 | 2043 | 2201 | 1490 | 2578 | 8647

<!--EndFragment-->
</body>
</html>


<html><body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Before changes |   |   |   |   |   |   |   |  
Orch seperated | 259 | 188 | 523 | 3108 | 3339 | 3643 | 3427 | 4895
After changes |   |   |   |   |   |   |   |  
Orch seperated | 285 | 171 | 612 | 2016 | 2288 | 1740 | 2646 | 8592

<!--EndFragment-->
</body>
</html>


### Type of change

- Optimization

### How risky is the change?

- Moderate